Release rk3566_rk3568 - 2023-12-28
Based on:
    e237b9d configs: debian: Add Debian12 (Bookworm) to support

+++ common/scripts/post-info.sh
@@ -2,18 +2,20 @@
 
 POST_ROOTFS_ONLY=1
 
-source "${POST_HELPER:-$(dirname "$(realpath "$0")")/../post-hooks/post-helper}"
+source "${RK_POST_HELPER:-$(dirname "$(realpath "$0")")/../post-hooks/post-helper}"
+
+[ "$RK_ROOTFS_DEBUG_INFO" ] || exit 0
 
 INFO_DIR="$TARGET_DIR/info"
 
-echo "Adding info dir..."
+message "Adding info dir..."
 
 rm -rf "$INFO_DIR"
 mkdir -p "$INFO_DIR"
 
-cd "$SDK_DIR"
+cd "$RK_SDK_DIR"
 
-yes | ${PYTHON3:-python3} .repo/repo/repo manifest -r \
+yes | python3 .repo/repo/repo manifest -r \
 	-o "$INFO_DIR/manifest.xml" &>/dev/null || true
 
 cat "$RK_CONFIG" | sed "s/\(PASSWORD=\)\".*\"/\1\"********\"/" > \
@@ -23,11 +25,11 @@ cp kernel/.config "$INFO_DIR/config-$RK_KERNEL_VERSION"
 cp kernel/System.map "$INFO_DIR/System.map-$RK_KERNEL_VERSION"
 
 EXTRA_FILES=" \
-	/etc/os-release /etc/fstab \
-	/proc/version /proc/cmdline /proc/kallsyms /proc/interrupts /proc/cpuinfo \
-	/proc/softirqs /proc/device-tree /proc/diskstats /proc/iomem \
-	/proc/meminfo /proc/partitions /proc/slabinfo \
-	/proc/rk_dmabuf /proc/rkcif-mipi-lvds /proc/rkisp0-vir0 \
+	/etc/os-release /etc/fstab /proc/config.gz \
+	/proc/cpuinfo /proc/version /proc/cmdline /proc/kallsyms \
+	/proc/interrupts /proc/softirqs /proc/device-tree /proc/diskstats \
+	/proc/iomem /proc/meminfo /proc/partitions /proc/slabinfo \
+	/proc/mpp_service /proc/rk_dmabuf /proc/rkcif-mipi-lvds /proc/rkisp0-vir0 \
 	/sys/kernel/debug/wakeup_sources /sys/kernel/debug/clk/clk_summary \
 	/sys/kernel/debug/gpio /sys/kernel/debug/pinctrl/ \
 	/sys/kernel/debug/dma_buf /sys/kernel/debug/dri \
